moved all test helpers to common.py, added protein_ids()
